+++ mods/iteminteractions/1.21.8/Common/src/main/java/fuzs/iteminteractions/api/v1/client/tooltip/ExpandableClientContentsTooltip.java
@@ -1,19 +1,21 @@
 package fuzs.iteminteractions.api.v1.client.tooltip;
 
 import fuzs.iteminteractions.impl.ItemInteractions;
-import fuzs.iteminteractions.impl.client.core.HeldActivationType;
+import fuzs.iteminteractions.impl.client.core.ActivationTypeProvider;
 import fuzs.iteminteractions.impl.config.ClientConfig;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
 
 public abstract class ExpandableClientContentsTooltip implements ClientTooltipComponent {
-    public static final String REVEAL_CONTENTS_TRANSLATION_KEY = "item.container.tooltip.revealContents";
+    public static final String REVEAL_CONTENTS_TRANSLATION_KEY = ItemInteractions.id("container")
+            .toLanguageKey(Registries.elementsDirPath(Registries.ITEM), "tooltip.reveal_contents");
 
     @Override
     public final int getHeight(Font font) {
-        if (!ItemInteractions.CONFIG.get(ClientConfig.class).revealContents.isActive()) {
+        if (!ItemInteractions.CONFIG.get(ClientConfig.class).visualItemContents.isActive()) {
             return 10;
         } else {
             return this.getExpandedHeight(font);
@@ -24,7 +26,7 @@ public abstract class ExpandableClientContentsTooltip implements ClientTooltipCo
 
     @Override
     public final int getWidth(Font font) {
-        HeldActivationType activation = ItemInteractions.CONFIG.get(ClientConfig.class).revealContents;
+        ActivationTypeProvider activation = ItemInteractions.CONFIG.get(ClientConfig.class).visualItemContents;
         if (!activation.isActive()) {
             Component component = activation.getComponent(REVEAL_CONTENTS_TRANSLATION_KEY);
             return font.width(component);
@@ -36,7 +38,7 @@ public abstract class ExpandableClientContentsTooltip implements ClientTooltipCo
 
     @Override
     public final void renderText(GuiGraphics guiGraphics, Font font, int x, int y) {
-        HeldActivationType activation = ItemInteractions.CONFIG.get(ClientConfig.class).revealContents;
+        ActivationTypeProvider activation = ItemInteractions.CONFIG.get(ClientConfig.class).visualItemContents;
         if (!activation.isActive()) {
             Component component = activation.getComponent(REVEAL_CONTENTS_TRANSLATION_KEY);
             guiGraphics.drawString(font, component, x, y, -1);
@@ -45,7 +47,7 @@ public abstract class ExpandableClientContentsTooltip implements ClientTooltipCo
 
     @Override
     public final void renderImage(Font font, int x, int y, int width, int height, GuiGraphics guiGraphics) {
-        if (!ItemInteractions.CONFIG.get(ClientConfig.class).revealContents.isActive()) return;
+        if (!ItemInteractions.CONFIG.get(ClientConfig.class).visualItemContents.isActive()) return;
         this.renderExpandedImage(font, x, y, guiGraphics);
     }