+++ mods/iteminteractions/1.21.8/Common/src/main/java/fuzs/iteminteractions/impl/data/client/ModLanguageProvider.java
@@ -2,9 +2,10 @@ package fuzs.iteminteractions.impl.data.client;
 
 import fuzs.iteminteractions.api.v1.client.tooltip.ExpandableClientContentsTooltip;
 import fuzs.iteminteractions.impl.ItemInteractions;
-import fuzs.iteminteractions.impl.client.core.HeldActivationType;
-import fuzs.iteminteractions.impl.client.core.KeyBackedActivationType;
-import fuzs.iteminteractions.impl.client.handler.KeyBindingTogglesHandler;
+import fuzs.iteminteractions.impl.client.core.ActivationTypeProvider;
+import fuzs.iteminteractions.impl.config.CarriedItemTooltips;
+import fuzs.iteminteractions.impl.config.SelectedItemTooltips;
+import fuzs.iteminteractions.impl.config.VisualItemContents;
 import fuzs.puzzleslib.api.client.data.v2.AbstractLanguageProvider;
 import fuzs.puzzleslib.api.data.v2.core.DataProviderContext;
 
@@ -17,11 +18,15 @@ public class ModLanguageProvider extends AbstractLanguageProvider {
     @Override
     public void addTranslations(TranslationBuilder builder) {
         builder.add(ExpandableClientContentsTooltip.REVEAL_CONTENTS_TRANSLATION_KEY, "%s %s to reveal contents");
-        builder.add(HeldActivationType.TOOLTIP_HOLD_TRANSLATION_KEY, "Hold");
-        builder.add(KeyBackedActivationType.KEY_TOOLTIP_PRESS_TRANSLATION, "Press");
-        builder.add(KeyBindingTogglesHandler.VISUAL_ITEM_CONTENTS_KEY.getKeyMapping(), "Toggle Visual Item Contents");
-        builder.add(KeyBindingTogglesHandler.SELECTED_ITEM_TOOLTIPS_KEY.getKeyMapping(), "Toggle Selected Item Tooltips");
-        builder.add(KeyBindingTogglesHandler.CARRIED_ITEM_TOOLTIPS_KEY.getKeyMapping(), "Toggle Carried Item Tooltips");
-        builder.add(KeyBackedActivationType.KEY_CATEGORY, ItemInteractions.MOD_NAME);
+        builder.add(ActivationTypeProvider.HOLD_COMPONENT, "Hold");
+        builder.add(ActivationTypeProvider.TOGGLE_COMPONENT, "Toggle");
+        builder.add(ActivationTypeProvider.SHIFT_COMPONENT, "Shift");
+        builder.add(ActivationTypeProvider.CONTROL_COMPONENT, "Control");
+        builder.add(ActivationTypeProvider.COMMAND_COMPONENT, "Command");
+        builder.add(ActivationTypeProvider.ALT_COMPONENT, "Alt");
+        builder.add(VisualItemContents.KEY_MAPPING, "Toggle Visual Item Contents");
+        builder.add(SelectedItemTooltips.KEY_MAPPING, "Toggle Selected Item Tooltips");
+        builder.add(CarriedItemTooltips.KEY_MAPPING, "Toggle Carried Item Tooltips");
+        builder.addKeyCategory(ItemInteractions.MOD_ID, ItemInteractions.MOD_NAME);
     }
 }