+++ mods/iteminteractions/1.21.8/CHANGELOG.md
@@ -1,16 +1,34 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v21.8.2-1.21.8] - 2025-07-21
+## [v21.8.3-1.21.8] - 2025-08-11
+
+### Changed
+
+- Update some string-based config options to use enum values instead
+
 ### Fixed
+
+- Fix carried item tooltips not showing
+
+## [v21.8.2-1.21.8] - 2025-07-21
+
+### Fixed
+
 - Fix filter for container items like shulker boxes not applying
 
 ## [v21.8.1-1.21.8] - 2025-07-20
+
 ### Fixed
+
 - Fix tooltips for items selected inside container items not showing
 
 ## [v21.8.0-1.21.8] - 2025-07-18
+
+### Changed
+
 - Update to Minecraft 1.21.8